First commit of 'automaton', the lua runtime
Acts as a client to the API, so this automaton doesn't actually
communicate with Firebase at all. It isn't skylink specific. It just
accepts a list of mounts (skylink for data, filesystem for source) and
then essentially exec()s the 'launch' routine from the source directory.

This runs ok right now but most probably has further bugs from the
copying of old JS.